+++ src/main/resources/application.properties
@@ -25,5 +25,5 @@ quarkus.http.auth.policy.role-policy1.roles-allowed=User
 quarkus.http.auth.permission.roles1.paths=/*
 quarkus.http.auth.permission.roles1.policy=role-policy1
 
-quarkus.http.auth.permission.permit1.paths=/secured/issue                                
+quarkus.http.auth.permission.permit1.paths=/data/security/issue
 quarkus.http.auth.permission.permit1.policy=permit
